+++ core/src/com/unciv/ui/screens/newgamescreen/FriendTable.kt
@@ -6,7 +6,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.badlogic.gdx.utils.Align
 import com.unciv.Constants
 import com.unciv.logic.multiplayer.FriendList
-import com.unciv.ui.components.WrappableLabel
+import com.unciv.ui.components.widgets.WrappableLabel
 import com.unciv.ui.components.extensions.pad
 
 class FriendTable(val friend: FriendList.Friend, width: Float, minHeight: Float)
