+++ core/src/com/unciv/ui/screens/newgamescreen/PlayerPickerTable.kt
@@ -18,7 +18,7 @@ import com.unciv.models.ruleset.nation.Nation
 import com.unciv.models.translations.tr
 import com.unciv.ui.components.input.KeyCharAndCode
 import com.unciv.ui.components.UncivTextField
-import com.unciv.ui.components.WrappableLabel
+import com.unciv.ui.components.widgets.WrappableLabel
 import com.unciv.ui.components.extensions.darken
 import com.unciv.ui.components.extensions.isEnabled
 import com.unciv.ui.components.input.keyShortcuts
